+++ db/migrate/20190416191650_create_videos.rb
@@ -1,0 +1,10 @@
+class CreateVideos < ActiveRecord::Migration[6.0]
+  def change
+    create_table :videos do |t|
+      t.string :name
+      t.text :embed_url
+
+      t.timestamps
+    end
+  end
+end