+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 8.0.0
 
 ### Enhancements
+- **Videos Added**
+  - Editors can embed videos for users to view
 - **Gem Changes**
   - Update to ruby 2.6.2
   - Update to rails 6.0.0.beta3